add ttnn tests
