+++ models/demos/deepseek_v3_b1/tests/unit_tests/test_mixed_tile.py
@@ -1,199 +1,139 @@
 # SPDX-FileCopyrightText: © 2025 Tenstorrent AI ULC
 # SPDX-License-Identifier: Apache-2.0
 
-"""CPU-only tests for mixed_tile assignment (no device needed)."""
+"""Tests for mixed_tile assignment using real ttnn BFP quantization."""
 
 from __future__ import annotations
 
-import numpy as np
-import pytest
 import torch
 
-from models.demos.deepseek_v3_b1.mixed_tile.assigner import MixedTileAssigner, MixedTileResult
-from models.demos.deepseek_v3_b1.mixed_tile.metrics import metric_is_good, metric_value, pearson_corr
-from models.demos.deepseek_v3_b1.mixed_tile.tile_utils import (
-    MIXED_TILE_FORMATS,
-    reconstruct_from_tiles,
-    reshape_to_2d_with_padding,
-    tile_metrics,
-)
+import ttnn
+from models.demos.deepseek_v3_b1.mixed_tile.assigner import MixedTileAssigner
+from models.demos.deepseek_v3_b1.mixed_tile.metrics import metric_value
+from models.demos.deepseek_v3_b1.mixed_tile.tile_utils import MIXED_TILE_BYTES_PER_ELEM, MIXED_TILE_FORMATS
 
 # ---------------------------------------------------------------------------
-# metrics tests
+# ttnn quantize-dequantize helpers
 # ---------------------------------------------------------------------------
 
-
-def test_pcc_identical():
-    a = np.random.randn(64).astype(np.float32)
-    assert pearson_corr(a, a) == pytest.approx(1.0)
-
-
-def test_pcc_zeros():
-    a = np.zeros(32, dtype=np.float32)
-    b = np.zeros(32, dtype=np.float32)
-    assert pearson_corr(a, b) == 1.0
+TTNN_DTYPE_MAP = {
+    "bf16": ttnn.bfloat16,
+    "bfp8": ttnn.bfloat8_b,
+    "bfp4": ttnn.bfloat4_b,
+}
 
 
-def test_pcc_different():
-    a = np.array([1.0, 2.0, 3.0], dtype=np.float32)
-    b = np.array([3.0, 2.0, 1.0], dtype=np.float32)
-    assert pearson_corr(a, b) == pytest.approx(-1.0)
-
-
-def test_metric_value_mae():
-    a = np.array([1.0, 2.0, 3.0], dtype=np.float32)
-    b = np.array([1.5, 2.5, 3.5], dtype=np.float32)
-    assert metric_value(a, b, "mae") == pytest.approx(0.5)
-
-
-def test_metric_value_atol():
-    a = np.array([1.0, 2.0, 3.0], dtype=np.float32)
-    b = np.array([1.0, 2.0, 5.0], dtype=np.float32)
-    assert metric_value(a, b, "atol") == pytest.approx(2.0)
-
-
-def test_metric_is_good_pcc():
-    assert metric_is_good(0.999, "pcc", 0.99)
-    assert not metric_is_good(0.98, "pcc", 0.99)
-
-
-def test_metric_is_good_mae():
-    assert metric_is_good(0.01, "mae", 0.05)
-    assert not metric_is_good(0.1, "mae", 0.05)
+def ttnn_quantize_fn(x: torch.Tensor, fmt: str) -> torch.Tensor:
+    """Quantize-dequantize round trip through ttnn."""
+    tt_dtype = TTNN_DTYPE_MAP[fmt]
+    tt_tensor = ttnn.from_torch(x, dtype=tt_dtype, layout=ttnn.TILE_LAYOUT)
+    return ttnn.to_torch(tt_tensor).to(dtype=torch.float32)
 
 
 # ---------------------------------------------------------------------------
-# tile_utils tests
+# Tests
 # ---------------------------------------------------------------------------
 
 
-def test_reshape_round_trip_2d():
-    x = np.random.randn(64, 64).astype(np.float32)
-    padded, shape_info, pad_info = reshape_to_2d_with_padding(x)
-    tile_hw = 32
-    tiles_h = pad_info[2] // tile_hw
-    tiles_w = pad_info[3] // tile_hw
-    tiles = padded.reshape(tiles_h, tile_hw, tiles_w, tile_hw).transpose(0, 2, 1, 3).reshape(-1, tile_hw, tile_hw)
-    reconstructed = reconstruct_from_tiles(tiles, shape_info, pad_info)
-    np.testing.assert_array_equal(reconstructed, x)
+def test_assignment_prefers_cheap_format_when_quality_is_met():
+    """On smooth data, bfp4 should be sufficient for most tiles at a reasonable threshold."""
+    torch.manual_seed(0)
+    # Smooth data (small values, low variance) → bfp4 should handle it fine
+    x = torch.randn(64, 64) * 0.01
+
+    assigner = MixedTileAssigner(metric="pcc", threshold=0.99, formats=["bfp8", "bfp4"])
+    result = assigner.assign(x, ttnn_quantize_fn)
+
+    assert result.tile_counts["bfp4"] > 0, "Expected at least some tiles assigned to bfp4"
+    total = result.tile_counts["bfp4"] + result.tile_counts["bfp8"]
+    assert total == 4  # 64x64 = 2x2 tiles
 
 
-def test_reshape_round_trip_1d():
-    x = np.random.randn(100).astype(np.float32)
-    padded, shape_info, pad_info = reshape_to_2d_with_padding(x)
-    tile_hw = 32
-    tiles_h = pad_info[2] // tile_hw
-    tiles_w = pad_info[3] // tile_hw
-    tiles = padded.reshape(tiles_h, tile_hw, tiles_w, tile_hw).transpose(0, 2, 1, 3).reshape(-1, tile_hw, tile_hw)
-    reconstructed = reconstruct_from_tiles(tiles, shape_info, pad_info)
-    np.testing.assert_array_equal(reconstructed, x)
-
-
-def test_tile_metrics_pcc_identical():
-    tiles = np.random.randn(4, 32, 32).astype(np.float32)
-    scores = tile_metrics(tiles, tiles, "pcc")
-    np.testing.assert_allclose(scores, 1.0, atol=1e-6)
-
-
-# ---------------------------------------------------------------------------
-# assigner tests
-# ---------------------------------------------------------------------------
-
-
-def _noop_quantize(x: np.ndarray, fmt: str) -> np.ndarray:
-    """Identity quantizer — returns input unchanged."""
-    return x.copy()
-
-
-def _noisy_quantize(x: np.ndarray, fmt: str) -> np.ndarray:
-    """Adds format-dependent noise: more noise for cheaper formats."""
-    noise_scale = {"bf16": 0.0, "bfp8": 0.001, "bfp4": 0.01, "bfp2": 0.1}
-    rng = np.random.default_rng(42)
-    return x + rng.normal(0, noise_scale.get(fmt, 0.01), size=x.shape).astype(np.float32)
-
-
-def test_assigner_noop_all_cheap():
-    """With identity quantizer, all tiles should get the cheapest format."""
-    x = np.random.randn(64, 64).astype(np.float32)
-    assigner = MixedTileAssigner(metric="pcc", threshold=0.999, formats=["bfp8", "bfp4"])
-    result = assigner.assign(x, _noop_quantize)
-
-    assert isinstance(result, MixedTileResult)
-    assert result.assignment.shape == (2, 2)
-    # All tiles should be bfp4 (cheapest) since identity quantizer gives PCC=1.0
-    fmt_to_idx = {fmt: idx for idx, fmt in enumerate(MIXED_TILE_FORMATS)}
-    assert np.all(result.assignment == fmt_to_idx["bfp4"])
-    assert result.tile_counts["bfp4"] == 4
-    assert result.tile_counts["bfp8"] == 0
-
-
-def test_assigner_noisy_mixed():
-    """With noisy quantizer and tight threshold, some tiles should need higher precision."""
-    rng = np.random.default_rng(123)
-    x = rng.normal(0, 1, size=(64, 64)).astype(np.float32)
-    assigner = MixedTileAssigner(metric="pcc", threshold=0.9999, formats=["bfp8", "bfp4"])
-    result = assigner.assign(x, _noisy_quantize)
-
-    assert result.assignment.shape == (2, 2)
-    # bfp4 adds more noise, so with a very tight threshold some tiles may need bfp8
-    total_tiles = result.tile_counts["bfp8"] + result.tile_counts["bfp4"]
-    assert total_tiles == 4
-
-
-def test_assigner_empty_tensor():
-    x = np.array([], dtype=np.float32)
-    assigner = MixedTileAssigner()
-    result = assigner.assign(x, _noop_quantize)
-    assert result.total_bytes == 0.0
-
-
-def test_assigner_invalid_metric():
-    with pytest.raises(ValueError, match="Unsupported metric"):
-        MixedTileAssigner(metric="invalid")
-
-
-def test_assigner_invalid_format():
-    with pytest.raises(ValueError, match="Unsupported format"):
-        MixedTileAssigner(formats=["fp16"])
-
-
-def test_assigner_result_shape_preserved():
-    """The quantized output should have the same shape as the input."""
-    x = np.random.randn(96, 128).astype(np.float32)
-    assigner = MixedTileAssigner(formats=["bfp8", "bfp4"])
-    result = assigner.assign(x, _noop_quantize)
-    assert result.quantized.shape == x.shape
-
-
-# ---------------------------------------------------------------------------
-# torch tensor tests
-# ---------------------------------------------------------------------------
-
-
-def _torch_noop_quantize(x: torch.Tensor, fmt: str) -> torch.Tensor:
-    return x.clone()
-
-
-def test_assigner_torch_input():
-    """Assigner should accept torch tensors and return torch quantized output."""
+def test_assignment_promotes_to_expensive_format_for_hard_tiles():
+    """Tiles with high dynamic range should get promoted to bfp8 under MAE metric."""
+    torch.manual_seed(0)
+    # Tile (0,0): mix of large and medium values — bfp4 (3-bit mantissa) will have
+    # significant absolute error compared to bfp8 (7-bit mantissa)
     x = torch.randn(64, 64)
+    x[:32, :32] = torch.linspace(-50, 50, 32 * 32).reshape(32, 32)
+    # Other tiles: near-zero, trivial for any format
+    x[32:, :] *= 0.001
+    x[:32, 32:] *= 0.001
+
+    # Use MAE with a tight threshold — bfp4 error on the linspace tile should exceed it
+    assigner = MixedTileAssigner(metric="mae", threshold=0.05, formats=["bfp8", "bfp4"])
+    result = assigner.assign(x, ttnn_quantize_fn)
+
+    fmt_to_idx = {fmt: idx for idx, fmt in enumerate(MIXED_TILE_FORMATS)}
+    # The linspace tile should need bfp8; the near-zero tiles should be fine with bfp4
+    assert (
+        result.assignment[0, 0] == fmt_to_idx["bfp8"]
+    ), f"Tile (0,0) with high dynamic range should be bfp8, got {MIXED_TILE_FORMATS[result.assignment[0, 0]]}"
+    # At least one other tile should be bfp4 (the near-zero tiles)
+    assert result.tile_counts["bfp4"] > 0, "Expected some tiles to remain bfp4"
+
+
+def test_quantized_output_meets_quality_threshold():
+    """The reconstructed tensor from mixed-tile should meet the global quality threshold."""
+    torch.manual_seed(42)
+    x = torch.randn(128, 128)
+
+    threshold = 0.999
+    assigner = MixedTileAssigner(metric="pcc", threshold=threshold, formats=["bfp8", "bfp4"])
+    result = assigner.assign(x, ttnn_quantize_fn)
+
+    pcc = metric_value(x.numpy(), result.quantized.numpy(), "pcc")
+    assert pcc >= threshold * 0.999, f"Global PCC {pcc:.6f} below threshold {threshold}"
+
+
+def test_mixed_tile_better_than_uniform_cheap():
+    """Mixed-tile should achieve equal or better quality than uniform bfp4."""
+    torch.manual_seed(7)
+    x = torch.randn(128, 128)
+    # Add some outlier tiles to make uniform bfp4 struggle
+    x[:32, :32] *= 50.0
+
     assigner = MixedTileAssigner(metric="pcc", threshold=0.999, formats=["bfp8", "bfp4"])
-    result = assigner.assign(x, _torch_noop_quantize)
+    result = assigner.assign(x, ttnn_quantize_fn)
 
-    assert isinstance(result.quantized, torch.Tensor)
-    assert result.quantized.shape == x.shape
-    assert result.assignment.shape == (2, 2)
+    uniform_bfp4 = ttnn_quantize_fn(x, "bfp4")
+    pcc_mixed = metric_value(x.numpy(), result.quantized.numpy(), "pcc")
+    pcc_uniform = metric_value(x.numpy(), uniform_bfp4.numpy(), "pcc")
+
+    assert (
+        pcc_mixed >= pcc_uniform
+    ), f"Mixed-tile PCC ({pcc_mixed:.6f}) should be >= uniform bfp4 PCC ({pcc_uniform:.6f})"
 
 
-def test_assigner_torch_numpy_same_assignment():
-    """Torch and numpy paths should produce identical assignments for the same data."""
-    rng = np.random.default_rng(42)
-    data = rng.normal(0, 1, size=(64, 64)).astype(np.float32)
+def test_mixed_tile_cheaper_than_uniform_expensive():
+    """Mixed-tile should use fewer bytes than uniform bfp8 when some tiles can be bfp4."""
+    torch.manual_seed(99)
+    x = torch.randn(128, 128)
 
-    assigner = MixedTileAssigner(metric="pcc", threshold=0.999, formats=["bfp8", "bfp4"])
+    assigner = MixedTileAssigner(metric="pcc", threshold=0.99, formats=["bfp8", "bfp4"])
+    result = assigner.assign(x, ttnn_quantize_fn)
 
-    result_np = assigner.assign(data, _noop_quantize)
-    result_torch = assigner.assign(torch.from_numpy(data), _torch_noop_quantize)
+    num_tiles = 16  # 128/32 * 128/32
+    uniform_bfp8_bytes = num_tiles * 1024 * MIXED_TILE_BYTES_PER_ELEM["bfp8"]
 
-    np.testing.assert_array_equal(result_np.assignment, result_torch.assignment)
-    np.testing.assert_allclose(result_np.quantized, result_torch.quantized.numpy(), atol=1e-6)
+    assert (
+        result.total_bytes <= uniform_bfp8_bytes
+    ), f"Mixed-tile bytes ({result.total_bytes:.0f}) should be <= uniform bfp8 ({uniform_bfp8_bytes:.0f})"
+
+
+def test_different_thresholds_different_assignments():
+    """A tighter threshold should assign more tiles to expensive formats."""
+    torch.manual_seed(123)
+    x = torch.randn(128, 128)
+    x[:32, :32] *= 20.0  # make one region harder
+
+    loose = MixedTileAssigner(metric="pcc", threshold=0.99, formats=["bfp8", "bfp4"])
+    tight = MixedTileAssigner(metric="pcc", threshold=0.9999, formats=["bfp8", "bfp4"])
+
+    result_loose = loose.assign(x, ttnn_quantize_fn)
+    result_tight = tight.assign(x, ttnn_quantize_fn)
+
+    assert result_tight.tile_counts["bfp8"] >= result_loose.tile_counts["bfp8"], (
+        f"Tight threshold should use >= bfp8 tiles: tight={result_tight.tile_counts['bfp8']}, "
+        f"loose={result_loose.tile_counts['bfp8']}"
+    )
